dashboard_requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
pillow>=10.0.0
python-multipart>=0.0.6